move varName to css plugin utils
